requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
aioredis==1.3.1           # via channels-redis
amqp==2.6.1               # via kombu
apipkg==1.5               # via execnet
asgiref==3.3.1            # via channels, channels-redis, daphne, django
async-timeout==3.0.1      # via aioredis
attrs==20.3.0             # via automat, flake8-bugbear, pytest, service-identity, twisted
autobahn==20.7.1          # via daphne
automat==20.2.0           # via twisted
beautifulsoup4-4.9.3
billiard==3.6.3.0         # via celery
celery==4.4.2             # via -r requirements/base.in, flower
certifi==2020.11.8        # via requests, sentry-sdk
cffi==1.14.4              # via cryptography, pynacl
channels-redis==2.4.2     # via thenewboston
channels==2.4.0           # via channels-redis, thenewboston
chardet==3.0.4            # via requests
constantly==15.1.0        # via twisted
coverage==5.3             # via pytest-cov
cryptography==3.2.1       # via autobahn, pyopenssl, service-identity
daphne==2.5.0             # via channels
django-cors-headers==3.5.0  # via thenewboston
django-debug-toolbar==3.1.1  # via thenewboston
django-filter==2.4.0      # via thenewboston
django-recaptcha-2.0.6
django-redis==4.12.1      # via thenewboston
django==3.1.3             # via channels, django-cors-headers, django-debug-toolbar, django-filter, django-redis, djangorestframework, thenewboston
djangorestframework==3.12.2  # via thenewboston
execnet==1.7.1            # via pytest-xdist
factory-boy==3.1.0        # via thenewboston
faker==5.0.0              # via factory-boy
fakeredis[lua]==1.4.4     # via thenewboston
flake8-bugbear==20.1.4    # via thenewboston
flake8-builtins==1.5.3    # via thenewboston
flake8-coding==1.3.2      # via thenewboston
flake8-docstrings==1.5.0  # via thenewboston
flake8-import-order==0.18.1  # via thenewboston
flake8-polyfill==1.0.2    # via thenewboston
flake8-quotes==3.2.0      # via thenewboston
flake8==3.8.4             # via flake8-bugbear, flake8-builtins, flake8-coding, flake8-docstrings, flake8-polyfill, flake8-quotes, thenewboston
flower==0.9.5             # via -r requirements/base.in
freezegun==1.0.0          # via thenewboston
hiredis==1.1.0            # via aioredis
humanize==3.1.0           # via flower
hyperlink==20.0.1         # via twisted
idna==2.10                # via hyperlink, requests, twisted
incremental==17.5.0       # via twisted
iniconfig==1.1.1          # via pytest
iptools==0.7.0            # via thenewboston
kombu==4.6.11             # via celery
lupa==1.9                 # via fakeredis
lxml-4.6.2
mccabe==0.6.1             # via flake8
msgpack==0.6.2            # via channels-redis
packaging==20.7           # via pytest
pluggy==0.13.1            # via pytest
prometheus-client==0.8.0  # via flower
psycopg2==2.8.6           # via thenewboston
ptable==0.9.2             # via thenewboston
py==1.9.0                 # via pytest, pytest-forked
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycodestyle==2.6.0        # via flake8, flake8-import-order, thenewboston
pycparser==2.20           # via cffi
pydocstyle==5.1.1         # via flake8-docstrings
pyflakes==2.2.0           # via flake8
pyhamcrest==2.0.2         # via twisted
pynacl==1.4.0             # via thenewboston
pyopenssl==20.0.0         # via twisted
pyparsing==2.4.7          # via packaging
pytest-asyncio==0.14.0    # via thenewboston
pytest-cov==2.10.1        # via thenewboston
pytest-django==4.1.0      # via thenewboston
pytest-forked==1.3.0      # via pytest-xdist
pytest-xdist==2.1.0       # via thenewboston
pytest==6.1.2             # via pytest-asyncio, pytest-cov, pytest-django, pytest-forked, pytest-xdist, thenewboston
python-dateutil==2.8.1    # via faker, freezegun
pytz==2020.5              # via celery, django, flower
redis==3.5.3              # via -r requirements/base.in, django-redis, fakeredis
requests-mock==1.8.0      # via thenewboston
requests==2.25.0          # via requests-mock, thenewboston
sentry-sdk==0.19.3        # via thenewboston
service-identity==18.1.0  # via twisted
six==1.15.0               # via automat, cryptography, fakeredis, pynacl, pyopenssl, python-dateutil, requests-mock
snowballstemmer==2.0.0    # via pydocstyle
sortedcontainers==2.3.0   # via fakeredis
soupsieve-2.1
sqlparse==0.4.1           # via django, django-debug-toolbar
tblib==1.7.0              # via -r requirements/base.in
text-unidecode==1.3       # via faker
thenewboston==0.0.30      # via -r requirements/base.in
toml==0.10.2              # via pytest
tornado==6.1              # via flower
twisted[tls]==20.3.0      # via daphne
txaio==20.4.1             # via autobahn
urllib3==1.26.2           # via requests, sentry-sdk
vine==1.3.0               # via amqp, celery
wheel==0.35.1             # via thenewboston
zope.interface==5.2.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
